Refactor HiveDataSource
The upcoming IcebergSplitReader will need to use fileHandleFactory_,
executor_, connectorQueryCtx_, etc because it needs to create another
HiveDataSource to read the delete files. This PR copies these required
fields to SplitReader. Moreover, most utility functions like makeScanspec
configureReaderOptions, testFilters are moved into the newly created
HiveConnectorUtil.h and cpp files, since they will be used by multiple
classes like the SplitReaders, HiveDataSource, and the upcoming Iceberg
DeleteFileReader.
