+++ velox/connectors/hive/SplitReader.h
@@ -16,39 +16,68 @@
 
 #pragma once
 
-#include "velox/dwio/common/Reader.h"
-#include "velox/type/Type.h"
+#include "velox/connectors/hive/FileHandle.h"
+#include "velox/dwio/common/Options.h"
+
+namespace facebook::velox {
+class BaseVector;
+class variant;
+using VectorPtr = std::shared_ptr<BaseVector>;
+} // namespace facebook::velox
+
+namespace facebook::velox::common {
+class MetadataFilter;
+class ScanSpec;
+} // namespace facebook::velox::common
+
+namespace facebook::velox::connector {
+class ConnectorQueryCtx;
+} // namespace facebook::velox::connector
 
 namespace facebook::velox::dwio::common {
-class BufferedInput;
+class Reader;
+class RowReader;
+class RuntimeStatistics;
+} // namespace facebook::velox::dwio::common
+
+namespace facebook::velox::memory {
+class MemoryPool;
 }
 
 namespace facebook::velox::connector::hive {
 
-constexpr const char* kPath = "$path";
-constexpr const char* kBucket = "$bucket";
-
 struct HiveConnectorSplit;
 class HiveTableHandle;
 class HiveColumnHandle;
+class HiveConfig;
 
 class SplitReader {
  public:
   static std::unique_ptr<SplitReader> create(
       std::shared_ptr<velox::connector::hive::HiveConnectorSplit> hiveSplit,
-      const RowTypePtr readerOutputType,
-      std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>&
-          partitionKeys,
+      std::shared_ptr<HiveTableHandle> hiveTableHandle,
       std::shared_ptr<common::ScanSpec> scanSpec,
-      memory::MemoryPool* pool);
+      const RowTypePtr readerOutputType,
+      std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>*
+          partitionKeys,
+      FileHandleFactory* fileHandleFactory,
+      folly::Executor* executor,
+      const ConnectorQueryCtx* connectorQueryCtx,
+      const std::shared_ptr<HiveConfig> hiveConfig,
+      std::shared_ptr<io::IoStatistics> ioStats);
 
   SplitReader(
       std::shared_ptr<velox::connector::hive::HiveConnectorSplit> hiveSplit,
-      const RowTypePtr readerOutputType,
-      std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>&
-          partitionKeys,
+      std::shared_ptr<HiveTableHandle> hiveTableHandle,
       std::shared_ptr<common::ScanSpec> scanSpec,
-      memory::MemoryPool* pool);
+      const RowTypePtr readerOutputType,
+      std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>*
+          partitionKeys,
+      FileHandleFactory* fileHandleFactory,
+      folly::Executor* executor,
+      const ConnectorQueryCtx* connectorQueryCtx,
+      const std::shared_ptr<HiveConfig> hiveConfig,
+      std::shared_ptr<io::IoStatistics> ioStats);
 
   virtual ~SplitReader() = default;
 
@@ -56,9 +85,6 @@ class SplitReader {
   /// do additional preparations before reading the split, e.g. Open delete
   /// files or log files, and add column adapatations for metadata columns
   virtual void prepareSplit(
-      const std::shared_ptr<HiveTableHandle>& hiveTableHandle,
-      const dwio::common::ReaderOptions& readerOptions,
-      std::unique_ptr<dwio::common::BufferedInput> baseFileInput,
       std::shared_ptr<common::MetadataFilter> metadataFilter,
       dwio::common::RuntimeStatistics& runtimeStats);
 
@@ -100,20 +126,23 @@ class SplitReader {
       const std::optional<std::string>& value) const;
 
   std::shared_ptr<HiveConnectorSplit> hiveSplit_;
-  RowTypePtr readerOutputType_;
-  std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>&
-      partitionKeys_;
+  std::shared_ptr<HiveTableHandle> hiveTableHandle_;
   std::shared_ptr<common::ScanSpec> scanSpec_;
-  memory::MemoryPool* pool_;
+  RowTypePtr readerOutputType_;
+  std::unordered_map<std::string, std::shared_ptr<HiveColumnHandle>>*
+      partitionKeys_;
+  memory::MemoryPool* const pool_;
   std::unique_ptr<dwio::common::Reader> baseReader_;
-  dwio::common::RowReaderOptions rowReaderOpts_;
   std::unique_ptr<dwio::common::RowReader> baseRowReader_;
+  FileHandleFactory* const fileHandleFactory_;
+  folly::Executor* const executor_;
+  const ConnectorQueryCtx* const connectorQueryCtx_;
+  const std::shared_ptr<HiveConfig> hiveConfig_;
+  std::shared_ptr<io::IoStatistics> ioStats_;
+  dwio::common::ReaderOptions baseReaderOpts_;
+  dwio::common::RowReaderOptions baseRowReaderOpts_;
 
  private:
-  void configureRowReaderOptions(
-      dwio::common::RowReaderOptions& options,
-      const RowTypePtr& rowType);
-
   bool emptySplit_;
 };
